updated website, final major change for the project.
Added icons that change based on the condition of the weather.

Tooltip for the icon, shows you a description of the weather

Changed map width to 80%

Better UI, added background colors to nav bar links

changed fonts

Footer looks better

Project is still not mobile friendly, this is due to the innate nature of canvas. They don't resize properly to changing sizes. And they get crunched up.